[MMCA-4787] Upgrade scala & library dependancies to latest version (#208)
* [MMCA-4787] library dependencies update from catalogue

* [MMCA-4787] Adding latest upgrade changes

* [MMCA-4787] Updating files according to recommendations

* [MMCA-4787] Adding lots of Scala changes

* [MMCA-4787] Removing unused params and fixing tests

* [MMCA-4787] fixing CustomsFinancialsConnectorSpec test

* [MMCA-4787] fixing PackageSpec test

* [MMCA-4787] fixing AuthorisedUserControllerSpec test

* [MMCA-4787] Fixing some compilation errors

* [MMCA-4787] Resolving PR comments

* [MMCA-4787] Updating build sbt file

* [MMCA-4787] Removing unused imports and updated build.sbt

* [MMCA-4787] Removing more unused

* [MMCA-4787] Removing whitespace

+++ app/repositories/AccountsRepository.scala
@@ -24,6 +24,7 @@ import play.api.Configuration
 import play.api.libs.json.{Format, Json, OFormat, Reads, Writes, __}
 import uk.gov.hmrc.mongo.play.PlayMongoComponent
 import uk.gov.hmrc.mongo.play.json.PlayMongoRepository
+import org.mongodb.scala.{SingleObservableFuture, ToSingleObservablePublisher}
 
 import java.time.{Instant, LocalDateTime, ZoneOffset}
 import java.util.concurrent.TimeUnit
@@ -45,7 +46,7 @@ class AccountsRepository @Inject()(
         IndexOptions()
           .name("accounts-last-updated-index")
           .unique(true)
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
       )
     )
   ) {
