+++ app/repositories/AuthoritiesRepository.scala
@@ -20,6 +20,7 @@ import com.mongodb.client.model.Indexes.ascending
 import models.domain.AuthoritiesWithId
 import org.mongodb.scala.model.Filters.equal
 import org.mongodb.scala.model.{IndexModel, IndexOptions, ReplaceOptions}
+import org.mongodb.scala.{SingleObservableFuture, ToSingleObservablePublisher}
 import play.api.Configuration
 import play.api.libs.json.{Json, OFormat, Reads, Writes}
 import uk.gov.hmrc.mongo.play.PlayMongoComponent
@@ -44,7 +45,7 @@ class AuthoritiesRepository @Inject()(
         ascending("lastUpdated"),
         IndexOptions().name("authorities-last-updated-index")
           .unique(true)
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInSeconds"), TimeUnit.SECONDS)
       )
     )
   ) {
